% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsm}
\alias{nlsm}
\title{Code Book for Ascii File
 Note: For more information, see the original article:
 David Card,
 "Using Geographic Variation in College Proximity to Estimate the Return to Schooling"
 NBER Working Paper 4832, August 1994}
\format{
A data frame with 3,613 rows and 52 variables:
\describe{
\item{id}{Sequential id runs from 1 to 5225}
\item{nearc2}{Grew up near 2-yr college}
\item{nearc4}{Grew up near 4-yr college}
\item{nearc4a}{Grew up near 4-yr public college}
\item{nearc4b}{Grew up near 4-yr priv college}
\item{ed76}{Education in 1976}
\item{ed66}{Education in 1966}
\item{age76}{Age in 1976}
\item{daded}{Dad's education missing=avg}
\item{nodaded}{1 if daded imputed}
\item{momed}{Mom's education}
\item{nomomed}{1 if momed imputed}
\item{weight}{nls weight for 1976 cross-section }
\item{momdad14}{if live with mom and dad age 14}
\item{sinmom14}{Lived with single mom age 14}
\item{step14}{Lived step parent age 14}
\item{reg661}{Dummy for region=1 in 1966}
\item{reg662}{Dummy for region=2 in 1966}
\item{reg663}{Dummy for region=3 in 1966}
\item{reg664}{Dummy for region=4 in 1966}
\item{reg665}{Dummy for region=5 in 1966}
\item{reg666}{Dummy for region=6 in 1966}
\item{reg667}{Dummy for region=7 in 1966}
\item{reg668}{Dummy for region=8 in 1966}
\item{reg669}{Dummy for region=9 in 1966}
\item{south66}{Lived in south in 1966}
\item{work76}{Worked in 1976}
\item{work78}{Worked in 1978}
\item{lwage76}{Log wage (outliers trimmed) 1976 }
\item{lwage78}{Log wage in 1978 outliers trimmed }
\item{famed}{Mom-dad education class 1-9}
\item{black}{1 if black}
\item{smsa76r}{In smsa in 1976}
\item{smsa78r}{In smsa in 1978}
\item{reg76r}{In south in 1976}
\item{reg78r}{In south in 1978}
\item{reg80r}{In south in 1980}
\item{smsa66r}{In smsa in 1966}
\item{wage76}{Raw wage cents per hour 1976}
\item{wage78}{Raw wage cents per hour 1978}
\item{wage80}{Raw wage cents per hour 1980}
\item{noint78}{1 if noninterview in 1978}
\item{noint80}{1 if noninterview in 1980}
\item{enroll76}{1 if enrolled in 1976}
\item{enroll78}{1 if enrolled in 1978}
\item{enroll80}{1 if enrolled in 1980}
\item{kww}{The kww score}
\item{iq}{A normed iq score}
\item{marsta76}{mar status in 1976 1=married, sp. present}
\item{marsta78}{mar status in 1978 1=married, sp. present}
\item{marsta80}{mar status in 1980 1=married, sp. present}
\item{libcrd14}{1 if lib card in home age 14}
}
}
\usage{
nlsm
}
\description{
This article is published (with the same title and identical tables)
in  Aspects of Labour Market Behaviour: Essays in Honour of John Vandekamp"
edited by Louis N. Christofides, E. Kenneth Grant, and Robert Swidinsky
Toronto: University of Toronto Press, 1995.
}
\details{
The file contains 3,613 observations on men in 1976 cross-section
of NLS young men (original NLS cohort)
}
\keyword{datasets}
