% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlsy97}
\alias{nlsy97}
\title{Highly processed version of NLS97 data.}
\format{
A data frame with 410 rows and 46 variables:
\describe{
  \item{lftwage}{Log of wages if positive and full-time}
  \item{black}{Indicator for black}
  \item{age}{Age}
  \item{age2}{Age squared}
  \item{msa}{Metropolitan area}
  \item{urban}{Indicator for urban}
  \item{south}{Indicator for south}
  \item{college}{Indicator for college}
  \item{female}{Indicator for female}
  \item{married}{Indicator for married}
  \item{children}{Indicator for children}
}
}
\source{
NLSY
}
\usage{
nlsy97
}
\description{
Highly processed version of NLS97 data.
}
\keyword{datasets}
